Cleanup includes, remove unnecessary transitive dependencies
Right now basically everything includes everything else.
This slows down compilation, because every small header file change will cause
a recompilation of most of the project. I've removed (most) of the unnecessary
dependencies with this PR, but there are some design problems - for example
Core.h and Utils.h files that are a thrash cans for random snippets of code
used by everything.