Add a nocheck modifier for index commend (#77)
* Add a nocheck modifier for index commend

Sometimes we want to index files *without* a safety belt that index
provides. One of the reasons ot do this is when the list of files doesn't
fit into RAM (or storing it there is too memory expensive to be useful).

In this case, curated list of files can be provided to the db, and `nocheck`
modifier tells the db that the file list is guaranteed to be unique.