Rework compacting. Respect RAM more.
Right now we don't care how many datasets we want to compact at once.
Ursadb believes it can handle it. Ursadb believes in itself very hard,
it doesn't change the fact that with few hundred datasets at once
compacting gets pretty hard.

This change ensures that the database won't try to compact more than
a hardcoded number of datasets, and also it won't try to compact
more than one set of datasets (heh) at once. Finally, it tries to check
for the possibility that datasets are locked, and ignore them in this
case.