Another specialisation for select - select with datasets. (#101)
Ok, why would anyone need this. This is intended to decrease speed to first
result in mquery. Instead of doing one big query, mquery will query datasets
one by one.

Now I can hear you asking: WHY would the logic for this be in mquery? Isn't
async querying a pretty generic use-case? So I agree, but implementing this
in mquery opens a can of worms that I'm not ready to handle, for example:

- Database can change during the query (solution: persistent snapshots?)
- Database can't know when it's OK to stop streaming (for example, user
  cancelled a query, or yara can't catch up).
- It makes database performance even more opaque - is it idle? Or are there
  tasks busy in the background?

Because of this, I think that querying DSets one by one is a good solution for
now. The proper solution would be something like persistent snapshots +
continuable iterators... But yeah, why use a cannon when small PR like this
will do the trick.

Bonus points: we can run queries in parallel!