Docker improvements (#179)
* feat(docker): set make to use multicore

* feat(docker): change default user to ursa

From a security perspective it is not recommended to run docker
processes under the default user, which is root. Therefore a `ursa` user
is created for all ursadb processes to run as.

* chore(ci): set automatic image rebuild

In order for docker images to contain the latest security patches it is
important to rebuild docker images at least every month. Upstream images
gets updated frequently and by rebuilding it is ensured that the ursadb
image contains these fixes.

Co-authored-by: RoyK <roy.kokkelkoren@ncsc.nl>